# backend/docs/swagger.json
{
    "components": {
        "schemas": {
            "database.Project": {
                "properties": {
                    "color": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "order": {
                        "type": "integer"
                    },
                    "tasks": {
                        "items": {
                            "$ref": "#/components/schemas/database.Task"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "database.Task": {
                "properties": {
                    "completed_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "due": {
                        "type": "string"
                    },
                    "end_datetime": {
                        "type": "string"
                    },
                    "google_event_id": {
                        "type": "string"
                    },
                    "has_time": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "labels": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "order": {
                        "type": "integer"
                    },
                    "project": {
                        "$ref": "#/components/schemas/database.Project"
                    },
                    "project_id": {
                        "type": "integer"
                    },
                    "recurrence": {
                        "type": "string"
                    },
                    "reminders": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "start_datetime": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "main.CreateTaskRequest": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "due": {
                        "type": "string"
                    },
                    "end_datetime": {
                        "type": "string"
                    },
                    "has_time": {
                        "type": "boolean"
                    },
                    "labels": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "order": {
                        "type": "integer"
                    },
                    "project_id": {
                        "type": "integer"
                    },
                    "recurrence": {
                        "type": "string"
                    },
                    "reminders": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "start_datetime": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "required": [
                    "title"
                ],
                "type": "object"
            },
            "main.CreateTaskResponse": {
                "properties": {
                    "task": {
                        "$ref": "#/components/schemas/database.Task"
                    },
                    "warning": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "main.SyncResponse": {
                "properties": {
                    "deleted_project_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "deleted_task_ids": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "projects": {
                        "items": {
                            "$ref": "#/components/schemas/database.Project"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "sync_token": {
                        "type": "string"
                    },
                    "tasks": {
                        "items": {
                            "$ref": "#/components/schemas/database.Task"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "main.UpdateTaskRequest": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "due": {
                        "type": "string"
                    },
                    "end_datetime": {
                        "type": "string"
                    },
                    "has_time": {
                        "type": "boolean"
                    },
                    "labels": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "order": {
                        "type": "integer"
                    },
                    "project_id": {
                        "type": "integer"
                    },
                    "recurrence": {
                        "type": "string"
                    },
                    "reminders": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "start_datetime": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "main.UpdateTaskResponse": {
                "properties": {
                    "warning": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "info": {
        "description": "Task management REST API",
        "title": "Dimaist API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/sync": {
            "get": {
                "operationId": "sync_data",
                "parameters": [
                    {
                        "description": "RFC3339 timestamp for incremental sync",
                        "in": "query",
                        "name": "sync_token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.SyncResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Sync data with incremental updates",
                "tags": [
                    "sync"
                ]
            }
        },
        "/tasks": {
            "get": {
                "operationId": "list_tasks",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/database.Task"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "List all tasks",
                "tags": [
                    "tasks"
                ]
            },
            "post": {
                "operationId": "create_task",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/main.CreateTaskRequest",
                                        "summary": "task",
                                        "description": "Task to create"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Task to create",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.CreateTaskResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    }
                },
                "summary": "Create a new task",
                "tags": [
                    "tasks"
                ]
            }
        },
        "/tasks/{task_id}": {
            "delete": {
                "operationId": "delete_task",
                "parameters": [
                    {
                        "description": "Task ID",
                        "in": "path",
                        "name": "task_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Delete a task",
                "tags": [
                    "tasks"
                ]
            },
            "get": {
                "operationId": "get_task",
                "parameters": [
                    {
                        "description": "Task ID",
                        "in": "path",
                        "name": "task_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/database.Task"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Get a task by ID",
                "tags": [
                    "tasks"
                ]
            },
            "put": {
                "operationId": "update_task",
                "parameters": [
                    {
                        "description": "Task ID",
                        "in": "path",
                        "name": "task_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/main.UpdateTaskRequest",
                                        "summary": "task",
                                        "description": "Task fields to update"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Task fields to update",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/main.UpdateTaskResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Update a task",
                "tags": [
                    "tasks"
                ]
            }
        },
        "/tasks/{task_id}/complete": {
            "post": {
                "operationId": "complete_task",
                "parameters": [
                    {
                        "description": "Task ID",
                        "in": "path",
                        "name": "task_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "summary": "Mark a task as complete",
                "tags": [
                    "tasks"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:3000/"
        }
    ]
}